Links! Added linkable functionality using the History.
EZ for when we can use GWT Hyperlinks, but for the links in our generic HTML in TopicWidget we needed to do some
clicklistening / address bar parsing to get the token, since History.new() wasn't getting called.

Currently the link button only links to Cathedral though, need to change the LinkListener.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@108 b8d5060c-ad15-0410-9d10-96fc93d4f2a4