Serious Progress on SeeAlsos. ie it appears to work.
Changed around what I was doing when I realized that if Feynman has a SeeAlso of Physics & Stanfors, and Physics has a SeeAlso of Chemistry, I didn't want this
to be one big SeeAlso fest, (Feynman, Stanford, Chemisty,Physics) which is what it was going to be. Now there'll be two see also, one for Feynman and one for 
Physics. If you want to see who's see also's you're included in, you can use List<TopicIdent> topicDAO.getLinksTo()

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@134 b8d5060c-ad15-0410-9d10-96fc93d4f2a4