Well, you've done it again, you've wasted another perfectly good day by totally rehashing the database after you had an epihany that meta's and MetaValues don't really exist, they're just Associations.
Things now appear to be right back where they were wrt functionality. We've deleted two tables & now have now Maps to deal with, 
just sets. The DB joins are starting to look a little scary though. Unlimitted extension capability will currently lead to unlimitted JOINs. Not sure 
if there's even a hope of getting out of that fix. Anyway, the queries are fast enough for now, and there is a certain beauty to it. 


git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@137 b8d5060c-ad15-0410-9d10-96fc93d4f2a4