WARN. possibly an invalid commit. started messing with occurrences and didn't finish.
Real fixes in this commit are wrt Subjects & Topics. Decided that Subjects shoudl not have islands, instead we should fake it and 
create a Tag for each Subject. On the fly in the GUI though. not as a domain required thing.

Also figured out that GWT IsSerializable was the cause of losing our HippoException messages. 

Some starter Facebook code. They say they're gonna rework the whole login system on Nov 6th though, so I'm gonna wait for that
to really dive in.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@156 b8d5060c-ad15-0410-9d10-96fc93d4f2a4