Added some feedback to the LoginWindow. TODO's here
Attempt at fixing the contentType parser in SubmitToHippo
Changed MessageService to use the DAO's directly instead of the Services, because those want Acegi users in the Security Context. 
An interesting question. Bit of an insecurity the way it's set up now.

Simplified the TopicDetails tab bar putting all references in the same tab. 

Start of add/edit Weblinks.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@211 b8d5060c-ad15-0410-9d10-96fc93d4f2a4