Load speed cheats. Put an invisible IFRAME in the decorator for all pages that loads the gwt site. NOTE: if the main site is getting a lot of traffic, this is probably a bad thing!  but for now, we'll start loading/caching right from the getgo.
Also removed the scripts that lived in the regular webapp and moved them to com.aavu/ this should let them be cached an ensure that we
upgrade them in sync.

This stuff required making com.aavu.hippo/ accessible in Acegi, but I think that's ok. 
It's the services that we want to lock down, not the interface to those services.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@225 b8d5060c-ad15-0410-9d10-96fc93d4f2a4