Finally, a new solution to the GWT serialization problems. Just rewrote ServerSerializationStreamWriter as ServerSerializationStreamWriterWithHibernateEscaping
Tough to plug this in, so unfortunately I needed to munge in a RemoteServiceServletreplacement and a GWTSpringControllerReplacement.

Added a cool progress bar and fixed problems with "this script is taking a long time"

Figured out that the preloader IFRAME was causing our double loads, so took that out for now.

Improvements to the way Island banners are drawn. Who knew getOffsetWidth() existed!?! fantastic!

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@249 b8d5060c-ad15-0410-9d10-96fc93d4f2a4