Now using the GWT-Simile-Library instead of our homegrown. Modified it in the same way so that we can load JSON strings.
Gave Occurrences a connection to the topics that contain them so that we can show users which of tags exist for 
links. Added the ability to see this for links and to add new links to an occurrence, but there's a little problem when 
creating new links bc the ID prop isn't returned and we end up duplicating.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@271 b8d5060c-ad15-0410-9d10-96fc93d4f2a4