New AddLink module that we'll use for the plugin
Shares functionality with the regular GUI's addLink stuff.

Touched a huge number of files, because I moved the constants out of Manager into a new ConstHolder class.
This was necessary if we were going to use the other module, but not use Manager. The only change in many of these files
is the new Consts location.

This update  also has highRes island images that we scale down 4x.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@277 b8d5060c-ad15-0410-9d10-96fc93d4f2a4