odd committing problems. out of date scripts directory? I dunno. I'm trying to delete it.
Plugin Updates from Roger.

Removed all use of scriptaculous w its prototype dependencies to help load speed.

MapGadget is functional and is saving locations on the map. Not successfull so far in multiple members per association.

Beginning a redo of the 'ViewMember/glossary' to make it more of an 'Explorer' wher you can pick which tags you want to see
and then see BlogView, Timeline, BigMap, or A-Z glossary.



git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@290 b8d5060c-ad15-0410-9d10-96fc93d4f2a4