New images for the zoomer. Moderate progress on getting the rest of the Explorer up and running. Going to try not to do too much premature optimization in queries and will push sorting burden to the client.
Added this to .equals of AbstractTopic & AbstractOccurrence to help make set.add() set.remove() work more like you'd
assume. Without this, if you update an obj, then add it to a set, the old one is still there and you get
nonunique problem. This seems like the best of business/natural worlds to me. Any issues w/ it?
if(this.getId() != 0 && 
   this.getId() == castOther.getId()){
			 return true;
}

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@294 b8d5060c-ad15-0410-9d10-96fc93d4f2a4