OpenID commit!
Ok, pretty large commit. Also changed around the way that signups get processed. People should only have to hit the mailing list now if there are more users than 'maxUsers' a var in applicationContext.xml.

OpenID is based on the org.acegi sandbox openID sutff and the JanRain openID libs. NOTE this is our first use of the session! so right now we're one JVM only. This is supposed to be avoidable by doing an SQL store instead of an in-memory one. Also beware, there may be some JanRain threading issues... that's what the blogosphere says at least.


git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@346 b8d5060c-ad15-0410-9d10-96fc93d4f2a4