Fixed search problem for openID users. calling compassQuery.term("username",username) was barfing on usernames like "http://bob.myopenid.com"
Switched to using user.getID, which was a better idea anyway. Also fixed a weirdo problem indexing, WrongClassException that was showing up in the logs, but this was due to a deployment of DEV.war that was out of date. phew!

MessageService should now let you email openIDs by looking up the username after doing an openID normalization so bob.myopenid.com@hipcamp.com should try to resolve http://bob.myopenid.com/

NOTE: we're doing a bit of processing based on the presence or absence of a '.' in the username, which is a bit sketchy. It's in the validator, so hopefully nobody sneaks a no dot openID username past us.



git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@352 b8d5060c-ad15-0410-9d10-96fc93d4f2a4