Beginning of some massive changes to the structure. Going to play around with having just one module, called Interactive, which will be a dispatcher module for AddLink, 'HippoTest' now UserMind, and a new PublicBrowse app. The motivation is that we want easy re-use of all our wonderfull widgets. Why write this site in freemarker? when we've got the superior love of GWT?
Trick one, was that this changes where we want to load images from. See notes in ImageHolder and dispatcherServlet about some clear.cache.gif weirdness. In general the theme is that this makes hosted mode a little uglier, since it takes us away from keeping everything in the nice little public directory, but my hope is that that will be a small price to pay.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@378 b8d5060c-ad15-0410-9d10-96fc93d4f2a4