Deleted Tag.java (although it's actually still in there bc hibernate needs it until we change the discriminators)
This was a big design decision, but it come down to the fact that the concept of 'Tag' (or at least the way we were using it) gave us only one level of hierarchy, and I decided that really wasn't enough. We need hierarchical power of folders.

New scripts.html which will run the script to make all topics without tags go under the user's Root tag. This will also solve our 'island of missing toys' problem.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@380 b8d5060c-ad15-0410-9d10-96fc93d4f2a4