Added the ability to add new Topics to the Timeline. Added the ability to add new HippoDates.
Made the context menu abstract, extended by two classes, one for ContextMenus in the hierarchy display, one for the Timeline.

Reworked Topic creation in MindscapeManager to be a bit more logical and allow for some of the new looping creation schemes.

topicService.createNewIfNonExistant(Topic t, Topic parent) will now ENSURE that this parent relationship exists. Before this relationship was only created if the topic did NOT exist.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@443 b8d5060c-ad15-0410-9d10-96fc93d4f2a4